Create BasicRobot.java
This is the beginning of the standard program that will always live on the Benson robot from Oct 2014 onwards. It is used for development and testing programs by the programmer team.  It also always needs to be easy to drive and operate for when we want to demonstrate the robot.   

The other program that can be easily loaded is "Aerial assist 2014".